Rename season-details module to seasons
