+++ app/build.gradle.kts
@@ -62,8 +62,8 @@ dependencies {
     implementation(projects.data.profile.implementation)
     implementation(projects.data.similar.api)
     implementation(projects.data.similar.implementation)
-    implementation(projects.data.seasonDetails.api)
-    implementation(projects.data.seasonDetails.implementation)
+    implementation(projects.data.seasons.api)
+    implementation(projects.data.seasons.implementation)
     implementation(projects.data.shows.api)
     implementation(projects.data.shows.implementation)
     implementation(projects.data.trailers.api)
