--- data/season-details/testing/build.gradle.kts
@@ -1,23 +1,0 @@
-plugins {
-    id("tvmaniac.kmm.library")
-}
-
-
-kotlin {
-    android()
-    ios()
-
-    sourceSets {
-        sourceSets["commonMain"].dependencies {
-            implementation(projects.core.database)
-            implementation(projects.core.util)
-            implementation(projects.data.seasonDetails.api)
-
-            implementation(libs.coroutines.core)
-        }
-    }
-}
-
-android {
-    namespace = "com.thomaskioko.tvmaniac.episodes.testing"
-}
