+++ data/seasons/implementation/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/implementation/SeasonsComponent.kt
@@ -1,0 +1,15 @@
+package com.thomaskioko.tvmaniac.seasons.implementation
+
+import com.thomaskioko.tvmaniac.seasons.api.SeasonsDao
+import com.thomaskioko.tvmaniac.seasons.api.SeasonsRepository
+import me.tatarka.inject.annotations.Provides
+
+interface SeasonsComponent {
+
+    @Provides
+    fun provideSeasonsRepository(bind: SeasonsRepositoryImpl): SeasonsRepository =
+        bind
+
+    @Provides
+    fun provideSeasonsDao(bind: SeasonsDaoImpl): SeasonsDao = bind
+}
