+++ data/seasons/testing/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/testing/FakeSeasonsRepository.kt
@@ -1,0 +1,34 @@
+package com.thomaskioko.tvmaniac.seasons.testing
+
+import com.thomaskioko.tvmaniac.core.db.Seasons
+import com.thomaskioko.tvmaniac.core.db.SelectSeasonWithEpisodes
+import com.thomaskioko.tvmaniac.core.networkutil.Either
+import com.thomaskioko.tvmaniac.core.networkutil.Failure
+import com.thomaskioko.tvmaniac.seasons.api.SeasonsRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.flowOf
+
+class FakeSeasonsRepository : SeasonsRepository {
+
+    private var seasonsResult = flowOf<Either<Failure, List<Seasons>>>()
+
+    private var seasonEpisodesResult = flowOf<Either<Failure, List<SelectSeasonWithEpisodes>>>()
+
+    suspend fun setSeasonsResult(result: Either<Failure, List<Seasons>>) {
+        seasonsResult = flow { emit(result) }
+    }
+
+    suspend fun setSeasonDetails(result: Either<Failure, List<SelectSeasonWithEpisodes>>) {
+        seasonEpisodesResult = flow { emit(result) }
+    }
+
+    override fun observeSeasonsStream(traktId: Long): Flow<Either<Failure, List<Seasons>>> = seasonsResult
+
+    override fun observeSeasonDetails(traktId: Long): Flow<Either<Failure, List<SelectSeasonWithEpisodes>>> =
+        seasonEpisodesResult
+
+    override fun observeSeasonDetailsStream(
+        traktId: Long,
+    ): Flow<Either<Failure, List<SelectSeasonWithEpisodes>>> = seasonEpisodesResult
+}
