+++ data/seasons/testing/src/commonMain/kotlin/com/thomaskioko/tvmaniac/seasons/testing/MockData.kt
@@ -1,0 +1,24 @@
+package com.thomaskioko.tvmaniac.seasons.testing
+
+import com.thomaskioko.tvmaniac.core.db.SelectSeasonWithEpisodes
+
+val seasonDetails = listOf(
+    SelectSeasonWithEpisodes(
+        trakt_id = 84958,
+        tmdb_id = 849583,
+        title = "Loki",
+        overview = "After stealing the Tesseract in Avengers: Endgame, Loki lands before the Time Variance Authority.",
+        runtime = 45,
+        id = 12345,
+        season_id = 12343,
+        season_number = 0,
+        episode_count = 15,
+        name = "Season 01",
+        trakt_id_ = 12345,
+        title_ = "Some title",
+        ratings = 4.5,
+        episode_number = "01",
+        votes = 4958,
+        image_url = "/kEl2t3OhXc3Zb9FBh1AuYzRTgZp.jpg",
+    ),
+)
