+++ presentation/seasondetails/build.gradle.kts
@@ -10,12 +10,12 @@ kotlin {
     sourceSets {
         sourceSets["commonMain"].dependencies {
             implementation(projects.data.episodes.api)
-            implementation(projects.data.seasonDetails.api)
+            implementation(projects.data.seasons.api)
         }
 
         sourceSets["commonTest"].dependencies {
             implementation(projects.data.episodes.testing)
-            implementation(projects.data.seasonDetails.testing)
+            implementation(projects.data.seasons.testing)
         }
     }
 }
