+++ settings.gradle.kts
@@ -54,9 +54,9 @@ include(
     ":data:episodes:api",
     ":data:episodes:implementation",
     ":data:episodes:testing",
-    ":data:season-details:api",
-    ":data:season-details:implementation",
-    ":data:season-details:testing",
+    ":data:seasons:api",
+    ":data:seasons:implementation",
+    ":data:seasons:testing",
     ":data:similar:api",
     ":data:similar:implementation",
     ":data:similar:testing",
